requirements.txt
prettytable
xlsxwriter
DrissionPage
requests
python-whois